+++ upgrades/0.0.102.eks-1-21/go.mod
@@ -1,4 +1,4 @@
-module github.com/oslokommune/okctl-upgrade/upgrades/okctl-upgrade/upgrades/0.0.103.eks-1-21
+module github.com/oslokommune/okctl-upgrade/upgrades/okctl-upgrade/upgrades/0.0.102.eks-1-21
 
 go 1.18
