+++ upgrades/0.0.102.eks-1-21/pkg/kubectl/api.go
@@ -3,7 +3,7 @@ package kubectl
 import (
 	"bytes"
 	"fmt"
-	"github.com/oslokommune/okctl-upgrade/upgrades/okctl-upgrade/upgrades/0.0.103.eks-1-21/pkg/kubectl/version"
+	"github.com/oslokommune/okctl-upgrade/upgrades/okctl-upgrade/upgrades/0.0.102.eks-1-21/pkg/kubectl/version"
 	"os"
 	"os/exec"
 )
