+++ upgrades/0.0.102.eks-1-21/pkg/kubectl/version/api_test.go
@@ -2,7 +2,7 @@ package version_test
 
 import (
 	_ "embed"
-	"github.com/oslokommune/okctl-upgrade/upgrades/okctl-upgrade/upgrades/0.0.103.eks-1-21/pkg/kubectl/version"
+	"github.com/oslokommune/okctl-upgrade/upgrades/okctl-upgrade/upgrades/0.0.102.eks-1-21/pkg/kubectl/version"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
