+++ upgrades/0.0.102.eks-1-21/pkg/kubectl/version/testdata/kubectlVersion.yaml
